requirements.txt
streamlit>=1.35.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
reportlab>=4.0.0
pandas>=2.0.0
